chore: Modify moved files to new class/namespace
This modifies the Contains extension methods to move it to a new namespace.

Note this commit is separate from the rename operation in order to minimize the chance that Git loses file history tracking.
