+++ Code/Light.GuardClauses/FrameworkExtensions/StringExtensions.cs
@@ -1,0 +1,7 @@
+﻿namespace Light.GuardClauses.FrameworkExtensions;
+
+/// <summary>
+/// Provides extension methods for the <see cref="string" /> class.
+/// </summary>
+// ReSharper disable once RedundantTypeDeclarationBody -- required for Source Code Transformation
+public static partial class StringExtensions { }
